change-logic-watchhistories-reaction

+++ CineWorld.Services.ReactionAPI/Controllers/FavoriteAPIController.cs
@@ -5,6 +5,7 @@ using CineWorld.Services.ReactionAPI.Models.Dtos;
 using CineWorld.Services.ReactionAPI.Utilities;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace CineWorld.Services.ReactionAPI.Controllers
 {
@@ -22,14 +23,16 @@ namespace CineWorld.Services.ReactionAPI.Controllers
             response = new ResponseDto();
         }
         [HttpGet]
-        [Route("{movieId:int}/{userId}")]
+        [Route("CheckFavorite/{movieId:int}")]
         [Authorize(Roles = $"{SD.AdminRole},{SD.CustomerRole}")]
-        public ResponseDto IsFavorited(int movieId, string userId)
+        public ResponseDto CheckFavorite(int movieId)
         {
             try
             {
-
+                string userId = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
                 bool isFavorited = _db.UserFavorites.Any(p => p.MovieId == movieId && p.UserId == userId);
+
+                response.IsSuccess = true;
                 response.Result = isFavorited;
             }
             catch (Exception ex)
@@ -38,17 +41,22 @@ namespace CineWorld.Services.ReactionAPI.Controllers
                 response.Message = ex.Message;
             }
             return response;
-
         }
         [HttpPost]
+        [Route("AddFavorite")]
         [Authorize(Roles = $"{SD.AdminRole},{SD.CustomerRole}")]
-        public ResponseDto Post([FromBody] UserFavoritesDto favoritesDto)
+        public ResponseDto AddFavorite([FromBody] UserFavoritesDto favoritesDto)
         {
             try
             {
+                string userId = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value; //Không cần gửi userID từ FE
+                favoritesDto.UserId = userId; 
+
                 UserFavorites obj = _mapper.Map<UserFavorites>(favoritesDto);
                 _db.UserFavorites.Add(obj);
                 _db.SaveChanges();
+
+                response.IsSuccess = true;
                 response.Result = favoritesDto;
             }
             catch (Exception ex)
@@ -59,19 +67,27 @@ namespace CineWorld.Services.ReactionAPI.Controllers
             return response;
         }
         [HttpDelete]
-        [Route("{movieId:int}/{userId}")]
+        [Route("RemoveFavorite/{movieId:int}")]
         [Authorize(Roles = $"{SD.AdminRole},{SD.CustomerRole}")]
-        public ResponseDto DeleteFavorite(int movieId, string userId)
+        public ResponseDto RemoveFavorite(int movieId)
         {
             try
             {
+                string userId = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
+                var favorite = _db.UserFavorites.FirstOrDefault(p => p.MovieId == movieId && p.UserId == userId);
 
-                var favorite = _db.UserFavorites
-                                  .FirstOrDefault(p => p.MovieId == movieId && p.UserId == userId);
-
-
-                _db.UserFavorites.Remove(favorite);
-                _db.SaveChanges();
+                if (favorite != null)
+                {
+                    _db.UserFavorites.Remove(favorite);
+                    _db.SaveChanges();
+                    response.IsSuccess = true;
+                    response.Message = "Đã xóa khỏi danh sách yêu thích.";
+                }
+                else
+                {
+                    response.IsSuccess = false;
+                    response.Message = "Phim không có trong danh sách yêu thích.";
+                }
             }
             catch (Exception ex)
             {
